Fix rhea script - and macox related reads counter
The rhea script was calling the main directly, skipping the argparse
which made the tool failed. Moving the arparse into the main fixed this.

Swap grep, zcat and wc -l in amplicon utils get_reads_count for pyfastx.
I had to change this as MacOS zcat doesn't seem to work the same as in
Linux (GNU vs BSD core utils). Initially swapped with biopython but
it was super slow, so I decided to use pyfastx. Now is as fast as
gzip, zcat and wc -l - and it handles gzip files automatically making
the code easier to read :).

I suspect we can get rid of biopython.