Study summary generator implementation improvements (#45)
* study-summary-generator should make summaries of ==1 run.
This is for consistency, so that we can reliably use
study summaries as a primary data fetch method without
users having to handle edge cases for single-run studies.

* extends logging of pandera schema failures.
Adds a helper for validating a df with a pandera schema,
which also accepts a metadata arg so that it is easy
to include something like a TSV filename in any failure
logs. This makes it easier to decipher which of many
TSV files failed validation and why.

* commits missing test files

---------

Co-authored-by: Martín Beracochea <mbc@ebi.ac.uk>